requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
discord==1.0.1
discord.py==1.7.2
idna==3.2
multidict==5.1.0
psutil==5.8.0
python-decouple==3.4
requests==2.25.1
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wget==3.2
yarl==1.6.3
